requirements.txt
# Core dependencies for RAG system
openai>=1.12.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
chromadb>=0.4.22
python-dotenv>=1.0.0

# Document processing
pypdf>=4.0.0
tiktoken>=0.5.2

# Optional but recommended
numpy>=1.24.0
